Pull in new schemas created in old repo
We put the process of migrating to this repo on hold as we had no
solution for updating schemas without cutting an ACS release. Since then
changes have been made to the `schemas` repo which need pulling across.

These are the new schemas. Modifications will need to be made by hand as
I don't want to lose the fixups I made before by scripting a re-import.